Update Approver Manager to use predicates. Test out predicates
